Heads Up! Prep... Create an application that starts with a GET request to pull data from the following API:
https://dojo-recipes.herokuapp.com/celebrities/

Use this data to populate a simple Recycler View.
create a button to allow users to add a new celebrity.
create a horizontal Linear Layout that holds and Entry Text and a Button, which allow the user to enter a name. Use the entered name to search the existing celebrities once the button is clicked. If the celebrity is in our Recycler View, bring the user to a new activity, that will allow the user to update or delete the selected celebrity.
Bonus:
Make use of coroutines to fetch API data in the background.